refactor(cli): split serve manager into pure ServeEngine + tokio serve layer (#99)
serve/manager.rs (1756 lines) tightly coupled the pure simulation
orchestration logic (state transitions, snapshotting, history flush,
boundary reset, dynamic add, the terminated-event ring) with the tokio
wiring (broadcast/oneshot channels, wall-clock pacing, the stream-io
bridge). That made the failure modes untestable without spinning up the
whole WebSocket stack, so they were only exercised by the heavy E2E.

Extract a pure `ServeEngine` (serve/engine.rs) that owns the simulation
state machine with no tokio, no channels, and no pacing:

- Outbound traffic leaves the engine as return values — pre-serialized
  broadcast strings on EngineInit/StepOutput/AddOutput and the paced
  HistoryState list on StepOutput::states — instead of `tx.send(...)`.
- stream-io bytes are pumped through an injected `StreamIo` sink/source;
  the serve layer provides a `BridgeStreamIo` adapter over the live
  endpoints, tests use a null/fake.
- `paused` is now serve-layer state (pause just stops calling step_chunk;
  it never changed physics), so the engine has no `paused` field.
- `propagate_to(t).unwrap()` becomes `?`: an integration error now halts
  gracefully like a controller fault instead of panicking the blocking
  task. This is the failure mode the engine unit tests can finally cover.
- The constructor's only input stays `Arc<SimParams>` so issue #98's
  future `SimulationPlan` is a one-line swap that never leaks into the
  serve loop or the connection layer.

manager.rs is now the thin tokio serve layer: it owns the
idle/running/paused state machine, mpsc/oneshot translation, broadcast,
spawn_blocking, wall-clock pacing, and the bridge lifecycle. The
SimCommand / SimStatusResponse contract is unchanged, so connection.rs
and mod.rs need no behavioural changes.

Add engine unit tests that need no runtime: construction + initial
Info/State emission, step_chunk time advance, atmospheric-entry
termination + replay ring, and the dynamic-add / construction-time
validation rejections (mixed attitude, streams without a controller,
attitude/streamed dynamic add). The moved helper tests (uniform_tick,
terminated ring, invalid_path_segment, body_names_for) keep passing.

Behaviour is preserved; the existing serve E2E remain the cross-check.
